Fix multipart test random failure.
Changes error handling to print in logs for out of band server errors in the mini server.

Moves the route entry up to the mini server declaration. This will prevent duplicate route promise routing. Where an event in one route triggers the other.

Adds a 20ms delay before the boundary test is executed. This prevents a connection reset during changeover on fast testers.